Linux fixes.

+++ src/common/Lgi/GFileSelect.cpp
@@ -1947,7 +1947,7 @@ void GFileSelect::ClearTypes()
 	d->Types.DeleteObjects();
 }
 
-GFileType *GFileSelect::TypeAt(int n)
+GFileType *GFileSelect::TypeAt(ssize_t n)
 {
 	return d->Types.ItemAt(n);
 }
